first

+++ Back/Shared/DTOS/PropDTOs/UnitDto.cs
@@ -6,7 +6,7 @@ namespace Shared.DTOS
         public int? PropertyId { get; set; }
         public string? UnitNumber { get; set; }
         public int Floor { get; set; }
-        public decimal Area { get; set; }
+        public double Area { get; set; }
         public string UsageType { get; set; }
         public string? FinishingType { get; set; }
         public string UnitType { get; set; }
